# openapi/openapi.yaml
...
openapi: 3.1.0
...
components:
  schemas:
    CadenceRequest:
      type: object
      properties:
        type:
          type: string
        dayOfMonth:
          type: integer
          format: int32
        expression:
          type: string
    ClauseRequest:
      type: object
      properties:
        eq:
          type: string
        in:
          type: array
          items:
            type: string
        gte:
          type: string
        lte:
          type: string
    DisableRequest:
      type: object
      properties:
        reason:
          type: string
    EnableRequest:
      type: object
      properties:
        reason:
          type: string
        catchUp:
          type: boolean
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
...
    PreviewRequest:
      type: object
      properties:
        selection:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ClauseRequest"
    Problem:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
...
    PutRequest:
      type: object
      properties:
        version:
          type: integer
          format: int64
        cadence:
          $ref: "#/components/schemas/CadenceRequest"
        timezone:
          type: string
        selection:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ClauseRequest"
        egressTemplateId:
          type: string
...
    RunNowResponse:
      type: object
      properties:
        exportBatchId:
          type: string
        jobId:
          type: string
        nextDueAt:
          $ref: "#/components/schemas/Instant"
    ScheduleView:
      type: object
      properties:
        tenantId:
          type: string
        vendor:
          type: string
        enabled:
          type: boolean
        cadence:
          $ref: "#/components/schemas/CadenceRequest"
        timezone:
          type: string
        selection:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ClauseRequest"
        egressTemplateId:
          type: string
        nextDueAt:
          $ref: "#/components/schemas/Instant"
        lastBatchId:
          type: string
        lastRunAt:
          $ref: "#/components/schemas/Instant"
        disabledAt:
          $ref: "#/components/schemas/Instant"
        disabledReason:
          type: string
        version:
          type: integer
          format: int64
        createdAt:
          $ref: "#/components/schemas/Instant"
        updatedAt:
          $ref: "#/components/schemas/Instant"
...
  securitySchemes:
    platform-token:
      description: Auth0 platform access token. The tenant-id header selects the tenant.
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /v1/vendor-exports/schedules:
    get:
      summary: List export schedules
      description: Schedules of the tenant named by the tenant-id header.
      tags:
      - vendor-exports
      parameters:
      - name: tenantId
        in: query
        schema:
          type: string
      responses:
        "400":
          description: "Invalid request (SELECTION_INVALID, CADENCE_INVALID, TIMEZONE_INVALID,\
            \ INVALID_REQUEST, TEMPLATE_NOT_FOUND, TEMPLATE_WRONG_ENTITY, REASON_REQUIRED)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: No or invalid platform token
        "403":
          description: "Caller holds no role in the tenant, lacks the permission,\
            \ or named another tenant (TENANT_MISMATCH)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: No such schedule (SCHEDULE_NOT_FOUND)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: "Exists, stale version, already enabled or disabled, ambiguous\
            \ template (SCHEDULE_EXISTS, VERSION_STALE, ALREADY_ENABLED, ALREADY_DISABLED,\
            \ TEMPLATE_AMBIGUOUS, BATCH_EXISTS, SCHEDULE_DISABLED)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "503":
          description: "DAL or api-layer unavailable or not configured (DAL_UNAVAILABLE,\
            \ API_LAYER_UNAVAILABLE, API_LAYER_NOT_CONFIGURED)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
      security:
      - platform-token: []
  /v1/vendor-exports/schedules/{tenantId}/{vendor}:
    put:
      summary: Create or replace an export schedule
      description: "No version in the body: create (201), or 409 SCHEDULE_EXISTS.\
        \ With version: replace (200), or 409 VERSION_STALE. Create provisions the\
        \ tenant's vendor template through api-layer and sets enabled; replace never\
        \ changes enabled. Cadence or timezone changes recompute nextDueAt."
      tags:
      - vendor-exports
      parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
      - name: vendor
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutRequest"
        required: true
      responses:
        "400":
          description: "Invalid request (SELECTION_INVALID, CADENCE_INVALID, TIMEZONE_INVALID,\
            \ INVALID_REQUEST, TEMPLATE_NOT_FOUND, TEMPLATE_WRONG_ENTITY, REASON_REQUIRED)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: No or invalid platform token
        "403":
          description: "Caller holds no role in the tenant, lacks the permission,\
            \ or named another tenant (TENANT_MISMATCH)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: No such schedule (SCHEDULE_NOT_FOUND)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: "Exists, stale version, already enabled or disabled, ambiguous\
            \ template (SCHEDULE_EXISTS, VERSION_STALE, ALREADY_ENABLED, ALREADY_DISABLED,\
            \ TEMPLATE_AMBIGUOUS, BATCH_EXISTS, SCHEDULE_DISABLED)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "503":
          description: "DAL or api-layer unavailable or not configured (DAL_UNAVAILABLE,\
            \ API_LAYER_UNAVAILABLE, API_LAYER_NOT_CONFIGURED)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleView"
        "200":
          description: Replaced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleView"
      security:
      - platform-token: []
    get:
      summary: Read an export schedule
      tags:
      - vendor-exports
      parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
      - name: vendor
        in: path
        required: true
        schema:
          type: string
      responses:
        "400":
          description: "Invalid request (SELECTION_INVALID, CADENCE_INVALID, TIMEZONE_INVALID,\
            \ INVALID_REQUEST, TEMPLATE_NOT_FOUND, TEMPLATE_WRONG_ENTITY, REASON_REQUIRED)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: No or invalid platform token
        "403":
          description: "Caller holds no role in the tenant, lacks the permission,\
            \ or named another tenant (TENANT_MISMATCH)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: No such schedule (SCHEDULE_NOT_FOUND)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: "Exists, stale version, already enabled or disabled, ambiguous\
            \ template (SCHEDULE_EXISTS, VERSION_STALE, ALREADY_ENABLED, ALREADY_DISABLED,\
            \ TEMPLATE_AMBIGUOUS, BATCH_EXISTS, SCHEDULE_DISABLED)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "503":
          description: "DAL or api-layer unavailable or not configured (DAL_UNAVAILABLE,\
            \ API_LAYER_UNAVAILABLE, API_LAYER_NOT_CONFIGURED)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
      security:
      - platform-token: []
  /v1/vendor-exports/schedules/{tenantId}/{vendor}/disable:
    post:
      summary: Disable an export schedule
      description: The tick skips the schedule until enable. Reason is mandatory.
      tags:
      - vendor-exports
      parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
      - name: vendor
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisableRequest"
        required: true
      responses:
        "400":
          description: "Invalid request (SELECTION_INVALID, CADENCE_INVALID, TIMEZONE_INVALID,\
            \ INVALID_REQUEST, TEMPLATE_NOT_FOUND, TEMPLATE_WRONG_ENTITY, REASON_REQUIRED)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: No or invalid platform token
        "403":
          description: "Caller holds no role in the tenant, lacks the permission,\
            \ or named another tenant (TENANT_MISMATCH)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: No such schedule (SCHEDULE_NOT_FOUND)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: "Exists, stale version, already enabled or disabled, ambiguous\
            \ template (SCHEDULE_EXISTS, VERSION_STALE, ALREADY_ENABLED, ALREADY_DISABLED,\
            \ TEMPLATE_AMBIGUOUS, BATCH_EXISTS, SCHEDULE_DISABLED)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "503":
          description: "DAL or api-layer unavailable or not configured (DAL_UNAVAILABLE,\
            \ API_LAYER_UNAVAILABLE, API_LAYER_NOT_CONFIGURED)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
      security:
      - platform-token: []
  /v1/vendor-exports/schedules/{tenantId}/{vendor}/enable:
    post:
      summary: Enable an export schedule
      description: catchUp true sets nextDueAt to the most recent occurrence missed
        while disabled; false to the next future one.
      tags:
      - vendor-exports
      parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
      - name: vendor
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnableRequest"
        required: true
      responses:
        "400":
          description: "Invalid request (SELECTION_INVALID, CADENCE_INVALID, TIMEZONE_INVALID,\
            \ INVALID_REQUEST, TEMPLATE_NOT_FOUND, TEMPLATE_WRONG_ENTITY, REASON_REQUIRED)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: No or invalid platform token
        "403":
          description: "Caller holds no role in the tenant, lacks the permission,\
            \ or named another tenant (TENANT_MISMATCH)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: No such schedule (SCHEDULE_NOT_FOUND)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: "Exists, stale version, already enabled or disabled, ambiguous\
            \ template (SCHEDULE_EXISTS, VERSION_STALE, ALREADY_ENABLED, ALREADY_DISABLED,\
            \ TEMPLATE_AMBIGUOUS, BATCH_EXISTS, SCHEDULE_DISABLED)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "503":
          description: "DAL or api-layer unavailable or not configured (DAL_UNAVAILABLE,\
            \ API_LAYER_UNAVAILABLE, API_LAYER_NOT_CONFIGURED)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
      security:
      - platform-token: []
  /v1/vendor-exports/schedules/{tenantId}/{vendor}/events:
    get:
      summary: Read a schedule's audit trail
      description: "The schedule-scoped events (no batch id), newest first, at most\
        \ 200. A batch's own events are under /v1/vendor-exports/{exportBatchId}/events."
      tags:
      - vendor-exports
      parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
      - name: vendor
        in: path
        required: true
        schema:
          type: string
      responses:
        "400":
          description: "Invalid request (SELECTION_INVALID, CADENCE_INVALID, TIMEZONE_INVALID,\
            \ INVALID_REQUEST, TEMPLATE_NOT_FOUND, TEMPLATE_WRONG_ENTITY, REASON_REQUIRED)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: No or invalid platform token
        "403":
          description: "Caller holds no role in the tenant, lacks the permission,\
            \ or named another tenant (TENANT_MISMATCH)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: No such schedule (SCHEDULE_NOT_FOUND)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: "Exists, stale version, already enabled or disabled, ambiguous\
            \ template (SCHEDULE_EXISTS, VERSION_STALE, ALREADY_ENABLED, ALREADY_DISABLED,\
            \ TEMPLATE_AMBIGUOUS, BATCH_EXISTS, SCHEDULE_DISABLED)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "503":
          description: "DAL or api-layer unavailable or not configured (DAL_UNAVAILABLE,\
            \ API_LAYER_UNAVAILABLE, API_LAYER_NOT_CONFIGURED)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
      security:
      - platform-token: []
  /v1/vendor-exports/schedules/{tenantId}/{vendor}/preview:
    post:
      summary: Preview a selection
      description: Runs the selection against api-layer with page size 1 and returns
        totalCount.
      tags:
      - vendor-exports
      parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
      - name: vendor
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PreviewRequest"
        required: true
      responses:
        "400":
          description: "Invalid request (SELECTION_INVALID, CADENCE_INVALID, TIMEZONE_INVALID,\
            \ INVALID_REQUEST, TEMPLATE_NOT_FOUND, TEMPLATE_WRONG_ENTITY, REASON_REQUIRED)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: No or invalid platform token
        "403":
          description: "Caller holds no role in the tenant, lacks the permission,\
            \ or named another tenant (TENANT_MISMATCH)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: No such schedule (SCHEDULE_NOT_FOUND)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: "Exists, stale version, already enabled or disabled, ambiguous\
            \ template (SCHEDULE_EXISTS, VERSION_STALE, ALREADY_ENABLED, ALREADY_DISABLED,\
            \ TEMPLATE_AMBIGUOUS, BATCH_EXISTS, SCHEDULE_DISABLED)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "503":
          description: "DAL or api-layer unavailable or not configured (DAL_UNAVAILABLE,\
            \ API_LAYER_UNAVAILABLE, API_LAYER_NOT_CONFIGURED)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
      security:
      - platform-token: []
  /v1/vendor-exports/schedules/{tenantId}/{vendor}/run-now:
    post:
      summary: Run an export schedule now
      description: "Creates the current period's batch as the tick would, advances\
        \ nextDueAt and enqueues the select job. 409 BATCH_EXISTS when the period\
        \ already has its batch, SCHEDULE_DISABLED when the schedule is disabled."
      tags:
      - vendor-exports
      parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
      - name: vendor
        in: path
        required: true
        schema:
          type: string
      responses:
        "400":
          description: "Invalid request (SELECTION_INVALID, CADENCE_INVALID, TIMEZONE_INVALID,\
            \ INVALID_REQUEST, TEMPLATE_NOT_FOUND, TEMPLATE_WRONG_ENTITY, REASON_REQUIRED)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: No or invalid platform token
        "403":
          description: "Caller holds no role in the tenant, lacks the permission,\
            \ or named another tenant (TENANT_MISMATCH)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: No such schedule (SCHEDULE_NOT_FOUND)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: "Exists, stale version, already enabled or disabled, ambiguous\
            \ template (SCHEDULE_EXISTS, VERSION_STALE, ALREADY_ENABLED, ALREADY_DISABLED,\
            \ TEMPLATE_AMBIGUOUS, BATCH_EXISTS, SCHEDULE_DISABLED)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "503":
          description: "DAL or api-layer unavailable or not configured (DAL_UNAVAILABLE,\
            \ API_LAYER_UNAVAILABLE, API_LAYER_NOT_CONFIGURED)"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "201":
          description: Batch created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunNowResponse"
      security:
      - platform-token: []
...
info:
  title: vendor-exchange-service
  version: 1.0.0
  description: "Directory accuracy vendor exchange: scheduled practitioner export\
    \ to an accuracy vendor, and later ingestion of the vendor's response."